bitcoingui: fix network weight, etc, not updating when many peers are lagging
If the median of the peers' block count is less than our block count,
things like network weight, etc, will not be updated. If we are connected
with slow to update peers, this can mean those update very seldom.
We now update those values when we're equal or above the median value.