stake prediction accuracy, and use display weight
Prediction of stake time gets complicated. Technically stake weight
begins to grow positive at 8 hours, but because of the stake modifier it
will not actually be eligible to stake until 8.8 days.  For display
purposes, this means weight should be 0 until 8.8 days. For stake
prediction times, this also means weight should be 0 until 8.8 days, but
after 8.8 it should be calculated using 8 hours as the point of time
that weight began to grow.

All stake time predictions are now synced. Since weight now accumulates
since 8 hours (although not displayed), the need to add an extra day of
weight to initial predictions is no longer needed and is removed.